+++ src/deion/worker/core/league/setGameAttributes.ts
@@ -1,8 +1,8 @@
 import { idb } from "../../db";
-import { g, toUI, initUILocalGames, local } from "../../util";
+import { g, toUI, helpers, initUILocalGames, local } from "../../util";
 import { unwrap, wrap } from "../../util/g";
 import type { GameAttributesLeague } from "../../../common/types";
-import { helpers } from "../../../common";
+import { finances } from "..";
 
 const updateMetaDifficulty = async (difficulty: number) => {
 	if (local.autoSave) {
@@ -50,6 +50,89 @@ const setGameAttributes = async (
 	for (const key of toUpdate) {
 		const value = wrap(g, key, gameAttributes[key]);
 		updated[key] = value;
+
+		if (key === "salaryCap") {
+			// Adjust budget items for inflation
+			if (
+				(g as any).salaryCap !== undefined &&
+				(g as any).season !== undefined &&
+				(g as any).userTids !== undefined
+			) {
+				const teams = await idb.cache.teams.getAll();
+				const teamSeasons = await idb.cache.teamSeasons.indexGetAll(
+					"teamSeasonsBySeasonTid",
+					[[g.get("season")], [g.get("season"), "Z"]],
+				);
+				const popRanks = helpers.getPopRanks(teamSeasons);
+
+				for (let i = 0; i < teams.length; i++) {
+					const t = teams[i];
+					console.log(t.region, t.name);
+					const popRank = popRanks[i];
+					if (popRank === undefined) {
+						continue;
+					}
+
+					if (g.get("userTids").includes(t.tid)) {
+						if (t.adjustForInflation !== false) {
+							// This won't perfectly keep pace, because a higher currentRank would be needed to maintain a higher rank, and a lower currentRank would be needed to maintain lower.
+							// Better - find out what popRank would be needed to obtain the current rank in each budget item
+							const currentRank = g.get("numTeams") / 2;
+
+							const factor =
+								helpers.defaultBudgetAmount(currentRank, value) /
+								helpers.defaultBudgetAmount(currentRank);
+							t.budget.coaching.amount =
+								Math.round((t.budget.coaching.amount * factor) / 10) * 10;
+							t.budget.facilities.amount =
+								Math.round((t.budget.facilities.amount * factor) / 10) * 10;
+							t.budget.health.amount =
+								Math.round((t.budget.health.amount * factor) / 10) * 10;
+							t.budget.scouting.amount =
+								Math.round((t.budget.scouting.amount * factor) / 10) * 10;
+
+							t.budget.ticketPrice.amount = parseFloat(
+								(
+									(t.budget.ticketPrice.amount *
+										helpers.defaultTicketPrice(currentRank, value)) /
+									helpers.defaultTicketPrice(currentRank)
+								).toFixed(2),
+							);
+						}
+					} else {
+						const defaultTicketPrice = helpers.defaultTicketPrice(
+							popRank,
+							value,
+						);
+						const defaultBudgetAmount = helpers.defaultBudgetAmount(
+							popRank,
+							value,
+						);
+
+						if (t.budget.ticketPrice.amount !== defaultTicketPrice) {
+							t.budget.ticketPrice.amount = defaultTicketPrice;
+						}
+
+						const keys: (keyof typeof t["budget"])[] = [
+							"scouting",
+							"coaching",
+							"health",
+							"facilities",
+						];
+
+						for (const key of keys) {
+							if (t.budget[key].amount !== defaultBudgetAmount) {
+								t.budget[key].amount = defaultBudgetAmount;
+							}
+						}
+					}
+
+					await idb.cache.teams.put(t);
+				}
+
+				await finances.updateRanks(["budget"]);
+			}
+		}
 
 		await idb.cache.gameAttributes.put({
 			key,