+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The dates are when development finished, not when the change was actually deploy
 
 Entries in bold are ones important enough that I would have put them in [the old changelog](./CHANGELOG-old.md).
 
+* 2020-05-16: Added "Auto adjust for inflation" option to the Team Finances page, which will keep revenue and expense settings in sync with salary cap changes. By default this setting is enabled.
+
 * 2020-05-14: Shows the team a player was selected from in the "Drafted" table of expansion drafts.
 
 * 2020-05-14: Historical rosters in BBGM include team abbreviations that change when a team moves, without breaking links to team pages from past seasons. Some parts of the UI still incorrectly display the current abbreviation rather than the historical one.